Started to work on two bindings for smartMDS/Ciseco and Open Energy monitor

+++ bundles/binding/org.openhab.binding.openenergymonitor/src/main/java/org/openhab/binding/openenergymonitor/protocol/OpenEnergyMonitorDataParser.java
@@ -122,17 +122,17 @@ public class OpenEnergyMonitorDataParser {
 	public static int toInt(byte[] data) {
 		// if (data == null || data.length != 4) return 0x0;
 
-		return (int) ((0xff & data[0]) << 24 | (0xff & data[1]) << 16
-				| (0xff & data[2]) << 8 | (0xff & data[3]) << 0);
+		return (0xff & data[0]) << 24 | (0xff & data[1]) << 16
+				| (0xff & data[2]) << 8 | (0xff & data[3]) << 0;
 	}
 
 	public static long toLong(byte[] data) {
 		// if (data == null || data.length != 8) return 0x0;
 
-		return (long) ((long) (0xff & data[0]) << 56
+		return (long) (0xff & data[0]) << 56
 				| (long) (0xff & data[1]) << 48 | (long) (0xff & data[2]) << 40
 				| (long) (0xff & data[3]) << 32 | (long) (0xff & data[4]) << 24
-				| (long) (0xff & data[5]) << 16 | (long) (0xff & data[6]) << 8 | (long) (0xff & data[7]) << 0);
+				| (long) (0xff & data[5]) << 16 | (long) (0xff & data[6]) << 8 | (long) (0xff & data[7]) << 0;
 	}
 
 	public static float toFloat(byte[] data) {
